Playlist generator by one genre and more genres methods:
- make PlaylistCredentials class and PlaylistCredentialsList class

- make generatePalylistByOneGenre method in PlylistRepository, Service And Controller

- make generatePalylistByMoreGenres method in PlylistRepository, Service And Controller

- make Constants interface to store constant Querries

+++ src/main/java/com/packt/app/genre/GenreRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface GenreRepository extends CrudRepository<Genre, Long> {
+public interface GenreRepository extends CrudRepository<Genre, Integer> {
 
     Genre findByName(String name);
