+++ src/main/java/com/packt/app/track/Track.java
@@ -9,6 +9,7 @@ import com.packt.app.artist.Artist;
 import com.packt.app.genre.Genre;
 
 import javax.persistence.*;
+import javax.persistence.criteria.CriteriaBuilder;
 
 @Entity
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -16,7 +17,7 @@ public class Track {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     private String title;
     private String link;
     private int duration;
@@ -57,11 +58,11 @@ public class Track {
         this.artist = artist;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
